import_role -> include_role
This change fixes a subtle bug where, if the same role was imported twice within a single play (e.g. to set up two different zenith client instances) then there could be some leakage of vars between role executions. This was observered, for example, when setting `zenith_forward_to_host: HOST_IP` in one case of importing the zenith client role and having that variable be picked up in another section of the play where the role is imported a second time without this var explicitly set. The expected behaviour here was that the var would be set to the role default, but instead the var was end up being shared with the other separate import of the role so that the `HOST_IP` value was set in all invocations of the role.